requirements.txt
uiautomator2>=3.0.0
lxml>=4.9.0
Pillow>=10.0.0
PyYAML>=6.0
